# Write dynamo benchmarks performance result to csv when throw exceptions