--- a/benchmarks/dynamo/ci_expected_accuracy/dynamic_cpu_inductor_torchbench_inference.csv
+++ b/benchmarks/dynamo/ci_expected_accuracy/dynamic_cpu_inductor_torchbench_inference.csv
@@ -74,30 +74,6 @@ detectron2_fasterrcnn_r_50_fpn,pass,46
 
 
 
-detectron2_fasterrcnn_r_101_c4,pass,42
-
-
-
-detectron2_fasterrcnn_r_101_dc5,pass,42
-
-
-
-detectron2_fasterrcnn_r_101_fpn,pass,46
-
-
-
-detectron2_fasterrcnn_r_50_c4,pass,42
-
-
-
-detectron2_fasterrcnn_r_50_dc5,pass,42
-
-
-
-detectron2_fasterrcnn_r_50_fpn,pass,46
-
-
-
 detectron2_fcos_r_50_fpn,pass,23